requirements.txt
langchain-openai
langchain
python-dotenv
langchain-community
langchain_core
streamlit
ipykernel
arxiv
wikipedia 
faiss-cpu
validators 
youtube_transcript_api
pytube
unstructured
ChromaDB
fastapi
uvicorn
requests
